LayoutTests/canvas/philip/tests/2d.transformation.translate.nonfinite-expected.txt

< [index] >
2d.transformation.translate.nonfinite
translate() with Infinity/NaN is ignored
References: 2d.nonfinite
Actual output:
Expected output:

Passed
